adding delete modal and create functional for admin

+++ icg-frontend/src/components/BoardAdmin.js
@@ -1,12 +1,34 @@
-import React, {useState, useEffect} from "react";
+import React, {useState, useEffect, useMemo} from "react";
 
 import axios from "axios";
 import {Link, useParams} from "react-router-dom";
+import {Button, Modal} from "react-bootstrap";
+import AuthService from "../services/AuthService";
 
 const BoardAdmin = () => {
-
     const [users, setUsers] = useState([]);
     const {id} = useParams();
+
+    const [show, setShow] = useState(false);
+    const [selectedUser, setSelectedUser] = useState({
+        id: 0,
+        name:''
+    });
+
+    const parseJwtToken = (token) =>  {
+        const base64Url = token.split('.')[1];
+        const base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/');
+        const jsonPayload = decodeURIComponent(window.atob(base64).split('').map(function(c) {
+            return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2);
+        }).join(''));
+        return JSON.parse(jsonPayload);
+    }
+
+    const user = AuthService.getCurrentUser();
+    const currentUser = parseJwtToken(user.token);
+
+    const handleClose = () => setShow(false);
+    const handleShow = () => setShow(true);
 
     useEffect(() => {
         loadUsers();
@@ -18,9 +40,31 @@ const BoardAdmin = () => {
     };
 
     const deleteUser = async (id) => {
+        if (users.find(u => u.id === id).email === currentUser.email) {
+            return;
+        }
+
         await axios.delete(`http://localhost:8080/api/user/${id}`);
         await loadUsers();
     };
+
+    const deleteModal = useMemo(() => {
+        if(selectedUser.id != 0)
+        return ( <Modal show={show} onHide={handleClose}>
+            <Modal.Header closeButton>
+                <Modal.Title>Are you sure?</Modal.Title>
+            </Modal.Header>
+            <Modal.Body><h6>Would you like to delete user <strong style={{color: "green"}}>{selectedUser.name}</strong> from the list ?</h6></Modal.Body>
+            <Modal.Footer>
+                <Button variant="secondary" onClick={handleClose}>
+                    Close
+                </Button>
+                <Button variant="primary" onClick={()=> {deleteUser(selectedUser.id); handleClose();}}>
+                    Delete
+                </Button>
+            </Modal.Footer>
+        </Modal>)
+    }, [selectedUser, show]);
 
     return (
 
@@ -39,28 +83,28 @@ const BoardAdmin = () => {
                 </button>
 
                 <Link className="btn btn-primary" to="/adduser">
-                    Add User
+                    Create User
                 </Link>
 
             </div>
 
 
             <div className="py-4">
-                <table className="table shadow">
+                <table className="table table-stripped">
                     <thead>
                     <tr>
-                        <th scope="col">#</th>
+                        <th scope="col" style={{paddingLeft:32}}>Id</th>
                         <th scope="col">Username</th>
                         <th scope="col">Email</th>
                         <th scope="col">Role</th>
-                        <th scope="col">Action</th>
+                        <th scope="col" style={{textAlign:"right"}}>Action</th>
                     </tr>
                     </thead>
                     <tbody>
-                    {users.map((user, index) => (
-                        <tr>
-                            <th scope="row" key={index}>
-                                {index + 1}
+                    {users.map(user =>
+                        <tr key={user.id}>
+                            <th scope="row" style={{paddingLeft:32}}>
+                                {user.id}
                             </th>
 
                             <td>{user.username}</td>
@@ -69,35 +113,44 @@ const BoardAdmin = () => {
                                 return <p> [{item.name}]</p>
                             })}</td>
                             <td>
-                                <Link
-                                    className="btn btn-primary mx-2"
-                                    to={`/viewuser/${user.id}`}
+                                <div className = "d-flex justify-content-end">
+                                    <Link
+                                        className="btn btn-primary mx-2"
+                                        to={`/viewuser/${user.id}`}
 
-                                >
-                                    View
-                                </Link>
-                                <Link
-                                    className="btn btn-primary mx-2"
-                                    to={{
-                                        pathname: `/edituser/${user.id}`,
-                                        state: { currentUser: user }
-                                    }}
-                                >
-                                    Edit
-                                </Link>
-                                <button
-                                    className="btn btn-danger mx-2"
-                                    onClick={() => deleteUser(user.id)}
-                                >
-                                    Delete
-                                </button>
+                                    >
+                                        View
+                                    </Link>
+                                    <Link
+                                        className="btn btn-primary mx-2"
+                                        to={{
+                                            pathname: `/edituser/${user.id}`,
+                                            state: { currentUser: user }
+                                        }}
+                                    >
+                                        Edit
+                                    </Link>
+
+
+                                    <button
+                                        className={`btn btn-danger mx-2 ${user.email === currentUser.email ? "d-none" : ""}`}
+                                        onClick={() =>{ setSelectedUser({...selectedUser, name: user.username, id: user.id}); handleShow()}}
+                                    >
+                                        Delete
+                                    </button>
+                                </div>
+
+
+
                             </td>
                         </tr>
-                    ))}
+                    )}
                     </tbody>
                 </table>
             </div>
+            {deleteModal}
         </div>
+
     );
 };
 export default BoardAdmin;
